# Scheduler-perf: reduce allocation of temporary resourceToValueMap objects on hotspot paths(use resourceToValueMapPool)