@novahe 2022-04-11
@denkensk  thanks for your reviewing

> Is this possible to reduce the efficiency of calculating score in parallel because of the race lock?

To answer this question, I write a benchmark to verify the performance of creating new map vs. using map object pool:

```golang
// ResourceName is the name identifying various resources in a ResourceList.
type ResourceName string

const (
	// CPU, in cores. (500m = .5 cores)
	ResourceCPU ResourceName = "cpu"
	// Memory, in bytes. (500Gi = 500GiB = 500 * 1024 * 1024 * 1024)
	ResourceMemory ResourceName = "memory"
	// Volume size, in bytes (e,g. 5Gi = 5GiB = 5 * 1024 * 1024 * 1024)
	ResourceStorage ResourceName = "storage"
	// Local ephemeral storage, in bytes. (500Gi = 500GiB = 500 * 1024 * 1024 * 1024)
	// The resource name for ResourceEphemeralStorage is alpha and it can change across releases.
	ResourceEphemeralStorage ResourceName = "ephemeral-storage"
)

type resourceToValueMap map[ResourceName]int64

func BenchmarkResourceToValueMapPool(b *testing.B) {
	resourceToValueMapPool := &sync.Pool{
		New: func() interface{} {
			return resourceToValueMap(make(map[ResourceName]int64, 4))
		},
	}
	for i := 0; i < b.N; i++ {
		requested := resourceToValueMapPool.Get().(resourceToValueMap)
		allocatable := resourceToValueMapPool.Get().(resourceToValueMap)
		for _, resource := range []ResourceName{ResourceCPU, ResourceMemory,
			ResourceEphemeralStorage, ResourceStorage} {
			requested[resource] = 1
			allocatable[resource] = 1
		}
		resourceToValueMapPool.Put(requested)
		resourceToValueMapPool.Put(allocatable)
	}
}

func BenchmarkNewResourceToValueMap(b *testing.B) {
	for i := 0; i < b.N; i++ {
		requested := make(resourceToValueMap)
		allocatable := make(resourceToValueMap)
		for _, resource := range []ResourceName{ResourceCPU, ResourceMemory,
			ResourceEphemeralStorage, ResourceStorage} {
			requested[resource] = 1
			allocatable[resource] = 1
		}
	}
}
```
benchmark result:
```
goos: windows
goarch: amd64
pkg: kubernetes
cpu: Intel(R) Core(TM) i7-8700 CPU @ 3.20GHz
BenchmarkResourceToValueMapPool
BenchmarkResourceToValueMapPool-12       9427284               122.8 ns/op
PASS

goos: windows
goarch: amd64
pkg: kubernetes
cpu: Intel(R) Core(TM) i7-8700 CPU @ 3.20GHz
BenchmarkNewResourceToValueMap
BenchmarkNewResourceToValueMap-12       12219411               105.9 ns/op
PASS
```
As a result, this modification doesn't improve performance, it only allocates less memory.
so I'll close this pr and submit a more valuable pr later.